pyproject.toml
[project]
name = "ai-newsletter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.12",
    "markdown>=3.10",
    "numpy>=2.3.5",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
    "tqdm>=4.67.1",
]
